Queen Class with move revolving methods
Queen class with constructor and two methods. validCoordinates method ensures a row/col entered is a valid location on the board, and that is follows the move capabilities of a queen (taking current location into account). checkMove method ensures all that validCoordinates does, but additionally checks to see if there are any other pieces in the way of this move.

These methods are essential to ensure players can only move pieces where allowed in traditional chess. Without these methods to ensure this, players could move anywhere on the board.