Updated Chess Piece class, Queen class, and new Pawn class
The check_move method was moved from the queen class to the Chess Piece class due as its implementation works for the majority of chess pieces. A Pawn class was created with the valid_move method ensure pawn pieces follow the right movement pattern

The pawn class is essential for the game as without it pawns would not move properly. Now pawns move as expected.